requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
exceptiongroup==1.1.3
fastapi==0.103.0
Flask==2.3.2
frozenlist==1.3.3
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
redis==4.6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
twilio==8.4.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.23.2
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.0
